Add files via upload
This file consist of the project alcohol Sensor. It is very crucial to know the adverse effects of drinking and driving.Thus,to combact this we have researched and found the limitations of the existing sensors and brought in a new technology to detect the concentration of alcohol in the human body with accuracy.